TEST: Move python functions to test_fortran.py
